// hello-world/src/generated/models/Cr1d7_healthmetricsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cr1d7_healthmetricsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cr1d7_healthmetricsstatecode = keyof typeof Cr1d7_healthmetricsstatecode;
export const Cr1d7_healthmetricsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cr1d7_healthmetricsstatuscode = keyof typeof Cr1d7_healthmetricsstatuscode;

export interface Cr1d7_healthmetricsBase {
  cr1d7_calorie?: string;
  cr1d7_carbs?: string;
  cr1d7_date?: string;
  cr1d7_deepsleeptime?: string;
  cr1d7_fat?: string;
  cr1d7_healthmetricid: string;
  cr1d7_hrv?: string;
  cr1d7_meditationminutes?: string;
  cr1d7_metricname: string;
  cr1d7_protein?: string;
  cr1d7_readinessscore?: string;
  cr1d7_remsleeptime?: string;
  cr1d7_sleepscore?: string;
  importsequencenumber?: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cr1d7_healthmetricsstatecode;
  statuscode?: Cr1d7_healthmetricsstatuscode;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
}

export interface Cr1d7_healthmetrics extends Cr1d7_healthmetricsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}